fix eslint errors
split AccountOutOfOffice to Form and Page files
